Add permanent storage
