+++ portfolio/pom.xml
@@ -28,6 +28,11 @@
         <artifactId>gson</artifactId>
         <version>2.8.6</version>
     </dependency>
+    <dependency>
+        <groupId>com.google.appengine</groupId>
+        <artifactId>appengine-api-1.0-sdk</artifactId>
+        <version>1.9.59</version>
+    </dependency>
   </dependencies>
 
   <build>
