+++ portfolio/src/main/webapp/index.html
@@ -105,7 +105,7 @@
 
                 <br>
                 <h3>Leave your comments here:</h3>
-                <form action="/comments" method="POST">
+                <form action="/new-comment" method="POST">
                     <label for="name">Your Name: </label>
                     <input type="text" name="name" style="width: 25%"><br><br>
                     <label for="commentText">Comment: </label>
